OCR improvements & bug fixes
Improved the OCR accuracy
Fixed a bug that was causing the app to crash due to a failure in converting an OCR result into an integer.
Fixed the signalling camera UI not being detected in specific signals.
Fixed trains running 2+ mins behind schedule not being accounted towards the total amount of trains in a SG zone